fix: check session in api

+++ src/routes/api/todo/delete/+server.ts
@@ -1,8 +1,12 @@
-import { supabase } from '$lib/supabase/client.js';
 import { json, error as svelteError } from '@sveltejs/kit';
 
 /** @type {import('./$types').RequestHandler} */
-export async function DELETE({ request }: { request: Request }) {
+export async function DELETE({ request, locals: { supabase, getSession } }) {
+	const session = await getSession();
+	if (!session) {
+		throw svelteError(401, { message: 'Unauthorized' });
+	}
+
 	const { targetId } = await request.json();
 	const { error, status } = await supabase.from('todo').delete().eq('id', targetId);
