+++ src/routes/api/todo/reset/+server.ts
@@ -2,7 +2,12 @@ import { supabase } from '$lib/supabase/client.js';
 import { json, error as svelteError } from '@sveltejs/kit';
 
 /** @type {import('./$types').RequestHandler} */
-export async function DELETE({ request }: { request: Request }) {
+export async function DELETE({ request, locals: { supabase, getSession } }) {
+	const session = await getSession();
+	if (!session) {
+		throw svelteError(401, { message: 'Unauthorized' });
+	}
+
 	const { targetStatus } = await request.json();
 	const { error, status } = await supabase.from('todo').delete().eq('status', targetStatus);
